chore(mda-start): organize functions
