+++ packages/mda-start/package.json
@@ -22,6 +22,7 @@
   "scripts": {
     "start": "node ./dist/esbuild/cli.js",
     "cli": "ts-node src/cli.ts",
+    "job": "ts-node src/job.ts",
     "lint": "eslint src/ --ext .js,.jsx,.ts,.tsx",
     "test": "jest",
     "clean": "rm -rf dist build package",
