+++ packages/mda-start/src/components/User/createOrUpdateMany.ts
@@ -1,36 +1,14 @@
 import { User, ZendeskUserCreationResponse } from "../../types";
-import fetch from "cross-fetch";
 import jobStatuses from "./jobStatuses";
 import logger from "../../logger";
+import zendeskRequest from "../zendeskRequest";
 
 const log = logger.child({ labels: { process: "createOrUpdateMany" } });
 
-const parameters = {
-  username: process.env.ZENDESK_API_USER || "",
-  token: process.env.ZENDESK_API_TOKEN || "",
-  remoteUri: process.env.ZENDESK_API_URL || ""
-}
-
-const getAuth = () => ({          
-  headers: {
-    Authorization: 'Basic ' + Buffer.from(parameters.username + "/token:" + parameters.token).toString('base64'),
-    'content-type': 'application/json'
-  }                     
-})
 
 const createOrUpdateMany = (users: User[]): Promise <ZendeskUserCreationResponse[]> => {
   
-  return fetch(`${parameters.remoteUri}users/create_or_update_many.json`, {
-    method: 'POST',
-    body: JSON.stringify({ users }),
-    ...getAuth()
-  })
-  .then((response) => { 
-    if(response.status === 200){
-      return response.json()
-    }
-    throw new Error(response.statusText);
-  })
+  return zendeskRequest(`users/create_or_update_many.json`, 'POST',JSON.stringify({ users }))
   .then((result) => {
     const jobID = result["job_status"].id;
